animated sword removed use of double

+++ src/main.cpp
@@ -21,13 +21,13 @@ int main(int argc, char* args[])
 
 		unsigned int g_CurrentTime = SDL_GetPerformanceCounter();;
 		unsigned int g_LastTime = 0;
-		double g_LastAnimStep = 0.0;
+		float g_LastAnimStep = 0.0;
 
 		while (g_App.IsRunning()) {
 			if (!g_App.IsPaused()) {
 				g_LastTime = g_CurrentTime;
 				g_CurrentTime = SDL_GetPerformanceCounter();
-				double deltaTime = ((g_CurrentTime - g_LastTime) / (double)SDL_GetPerformanceFrequency());
+				float deltaTime = ((g_CurrentTime - g_LastTime) / (float)SDL_GetPerformanceFrequency());
 
 				g_LastAnimStep > 0.125 ? g_App.UpdateAnimation(&g_LastAnimStep) : g_LastAnimStep += deltaTime;
